requirements.txt
beautifulsoup4
requests
python-Levenshtein